Apply review suggestions
* Move the textfield and button into their own component
* Rename production url to default url
* Rename textfield ref
