+++ source/lib/constants.ts
@@ -1,3 +1,3 @@
 export const GH_BASE_URL = 'https://github.com/StoDevX/AAO-React-Native'
 export const GH_NEW_ISSUE_URL = `${GH_BASE_URL}/issues/new`
-export const PRODUCTION_SERVER_URL = 'https://stolaf.api.frogpond.tech/v1'
+export const DEFAULT_URL = 'https://stolaf.api.frogpond.tech/v1'
