+++ source/views/settings/screens/overview/developer.tsx
@@ -1,28 +1,16 @@
 import * as Sentry from '@sentry/react-native'
 import * as React from 'react'
-import {Alert, StyleSheet} from 'react-native'
-import {
-	Section,
-	PushButtonCell,
-	CellTextField,
-	ButtonCell,
-} from '@frogpond/tableview'
+import {Alert} from 'react-native'
+import { Section, PushButtonCell } from '@frogpond/tableview'
 import type {NavigationScreenProp} from 'react-navigation'
 import {isDevMode} from '@frogpond/constants'
-import restart from 'react-native-restart'
-import * as storage from '../../../../lib/storage'
-import {PRODUCTION_SERVER_URL} from '../../../../lib/constants'
+import { ServerUrlSection } from './server-url'
 
 type Props = {
 	navigation: NavigationScreenProp<any>
 }
 
 export const DeveloperSection = ({navigation}: Props): React.ReactElement => {
-	const [serverAddress, setServerAddress] = React.useState('')
-	const _serverAddress = React.useRef<CellTextField>(null)
-
-	const [errorMessage, setErrorMessage] = React.useState('')
-
 	const onAPIButton = () => navigation.navigate('APITestView')
 	const onBonAppButton = () => navigation.navigate('BonAppPickerView')
 	const onDebugButton = () => navigation.navigate('DebugView')
@@ -47,31 +35,6 @@ export const DeveloperSection = ({navigation}: Props): React.ReactElement => {
 		}
 	}
 
-	const checkForValidity = (text: string) => {
-		const pattern = /^(http|https):\/\/[^ "]+$/u
-		const isUrlValid = pattern.test(text)
-		const isValid = isUrlValid || text.length === 0
-
-		isValid ? setErrorMessage('') : setErrorMessage('Waiting for a valid url…')
-	}
-
-	const handleOnChange = (text: string) => {
-		setServerAddress(text)
-		checkForValidity(text)
-	}
-
-	const refreshApp = () => {
-		storage.setServerAddress(serverAddress)
-		restart.Restart()
-	}
-
-	React.useMemo(async () => {
-		if (!serverAddress && !_serverAddress.current) {
-			const address = await storage.getServerAddress()
-			setServerAddress(address)
-		}
-	}, [serverAddress])
-
 	return (
 		<>
 			<Section header="DEVELOPER">
@@ -88,32 +51,7 @@ export const DeveloperSection = ({navigation}: Props): React.ReactElement => {
 				/>
 			</Section>
 
-			<Section
-				footer={
-					serverAddress ? '' : 'Empty means we will use the production server.'
-				}
-				header="SERVER"
-			>
-				<CellTextField
-					key={0}
-					_ref={_serverAddress}
-					onChangeText={handleOnChange}
-					placeholder={`${PRODUCTION_SERVER_URL}`}
-					value={serverAddress}
-				/>
-				<ButtonCell
-					disabled={errorMessage.length > 0}
-					onPress={refreshApp}
-					textStyle={styles.buttonCell}
-					title={errorMessage ? errorMessage : 'Update Server URL'}
-				/>
-			</Section>
+			<ServerUrlSection />
 		</>
 	)
 }
-
-const styles = StyleSheet.create({
-	buttonCell: {
-		textAlign: 'center',
-	},
-})
